fix bug

+++ new_stick.ino
@@ -54,7 +54,7 @@ void setup() {
 }
 
 void loop() {
-  const uint8_t* image = currect_image();
+  const uint16_t* image = currect_image();
   char d = 0;
   readSpi(REG_ACCEL_XOUT_H, (uint8_t*)&d, sizeof(d), CS);
   int line = ((int)d + 0x80) * IMG_WIDTH / 0x100;
@@ -65,7 +65,7 @@ void loop() {
   }
 }
 
-static void draw_pixels(const uint8_t* image, int line) {
+static void draw_pixels(const uint16_t* image, int line) {
   for (int i = 0; i < IMG_HEIGHT; ++i) {
     uint16_t c = pgm_read_word(image + line * IMG_HEIGHT + i);
     uint8_t r = (c & 0xF800) >> 8;
@@ -75,7 +75,7 @@ static void draw_pixels(const uint8_t* image, int line) {
   }
 }
 
-static const uint8_t* currect_image() {
+static const uint16_t* currect_image() {
   static int nimg = 0;
   static int btn0 = HIGH;
   int btn1 = digitalRead(BUTTON_PIN);
